+++ tests/unit_tests/test_internal_api.py
@@ -10,6 +10,7 @@ import pytest
 import requests
 import responses
 import wandb.errors
+from wandb.proto.wandb_internal_pb2 import ServerFeature
 import wandb.sdk.internal.internal_api
 import wandb.sdk.internal.progress
 from wandb.apis import internal
@@ -732,3 +733,106 @@ class TestUploadFileRetry:
             )
 
         assert handler.call_count == num_retries + 1
+    
+ENABLED_FEATURE_RESPONSE = {
+    "serverInfo": {
+        "features": [
+            {"name": "LARGE_FILENAMES", "isEnabled": True},
+            {"name": "ARTIFACT_TAGS", "isEnabled": False},
+        ]
+    }
+}
+
+
+@pytest.fixture
+def mock_client(mocker):
+    mock = mocker.patch("wandb.sdk.internal.internal_api.Client")
+    mock.return_value = mocker.Mock()
+    yield mock.return_value
+
+
+@pytest.fixture
+def mock_client_with_enabled_features(mock_client):
+    mock_client.execute.return_value = ENABLED_FEATURE_RESPONSE
+    yield mock_client
+
+
+NO_FEATURES_RESPONSE = {"serverInfo": {"features": []}}
+
+
+@pytest.fixture
+def mock_client_with_no_features(mock_client):
+    mock_client.execute.return_value = NO_FEATURES_RESPONSE
+    yield mock_client
+
+
+@pytest.fixture
+def mock_client_with_error_no_field(mock_client):
+    error_msg = 'Cannot query field "features" on type "ServerInfo".'
+    mock_client.execute.side_effect = Exception(error_msg)
+    yield mock_client
+
+
+@pytest.fixture
+def mock_client_with_random_error(mock_client):
+    error_msg = "Some random error"
+    mock_client.execute.side_effect = Exception(error_msg)
+    yield mock_client
+
+
+@pytest.mark.parametrize(
+    "fixture_name, feature, expected_result, expected_error",
+    [
+        # Test enabled features
+        (
+            "mock_client_with_enabled_features",
+            ServerFeature.LARGE_FILENAMES,
+            True,
+            False,
+        ),
+        # Test disabled features
+        (
+            "mock_client_with_enabled_features",
+            ServerFeature.ARTIFACT_TAGS,
+            False,
+            False,
+        ),
+        # Test features not in response
+        (
+            "mock_client_with_enabled_features",
+            ServerFeature.ARTIFACT_REGISTRY_SEARCH,
+            False,
+            False,
+        ),
+        # Test empty features list
+        ("mock_client_with_no_features", ServerFeature.LARGE_FILENAMES, False, False),
+        # Test server not supporting features
+        (
+            "mock_client_with_error_no_field",
+            ServerFeature.LARGE_FILENAMES,
+            False,
+            False,
+        ),
+        # Test other server errors
+        ("mock_client_with_random_error", ServerFeature.LARGE_FILENAMES, False, True),
+    ],
+)
+@pytest.mark.usefixtures("patch_apikey", "patch_prompt")
+def test_server_feature_checks(
+    request,
+    fixture_name,
+    feature,
+    expected_result,
+    expected_error,
+):
+    """Test check_server_feature with various scenarios."""
+    request.getfixturevalue(fixture_name)
+    api = internal.InternalApi()
+
+    if expected_error:
+        with pytest.raises(Exception, match="Some random error"):
+            api._check_server_feature_with_fallback(feature)
+    else:
+        result = api._check_server_feature_with_fallback(feature)
+        assert result == expected_result
+
